Conflicts:
	Builds/MLEngine_Debug.exe
	Builds/MLEngine_Debug.ilk
	Builds/MLEngine_Debug.pdb
	Builds/MLEngine_Release.exe
	Builds/MLEngine_Release.pdb
	MLEngine.v12.suo

+++ MLEngine/Code/Components/DebugComponent.cpp
@@ -10,7 +10,7 @@ void DebugComponent::Init()
 	std::cout << "Debug Component has been Inited!!" << std::endl;
 }
 
-void DebugComponent::Update(float dt)
+void DebugComponent::Update(double dt)
 {
 	std::cout << "Debug Component has been updated with delta t: " << dt << std::endl;
 }
